requirements.txt
disnake~=2.11.0
python-dotenv~=1.2.1